Restructure Banking System (#244)
Move transaction history components to system.banking.history package and banking operations to system.banking.operations package. 

Update API paths to match the new structure and improve documentation.
